pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xmutant-transformer"
version = "0.1.0"
description = "Adversarial testing for transformer models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "captum>=0.8.0",
    "datasets>=4.1.1",
    "matplotlib-venn>=1.1.2",
    "numpy>=1.26.4",
    "openai>=1.109.1",
    "pandas>=2.3.2",
    "protobuf>=6.32.1",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.2",
    "typer>=0.19.2",
]

[project.scripts]
xmutant = "xmutant_transformer.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["xmutant_transformer*"]
